requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.0
httpx>=0.27.0
pypdf>=4.0.0
python-docx>=1.1.0
python-multipart>=0.0.9
